Many updates, refactoring
